scripts/data_acquisition/gcf_build_panel/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.0
pandas>=2.0
pyarrow>=14.0
pyyaml>=6.0
